Save WIP
テスト用クラスの分離．
サブクラス読み込みテストの追加．
